+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 1.3.2
 
 ### 🐛 Bug Fix